refactor: move pool methods into pools folder
