+++ packages/vitest/src/node/core.ts
@@ -13,7 +13,7 @@ import { deepMerge, hasFailed, noop, slash, toArray } from '../utils'
 import { getCoverageProvider } from '../integrations/coverage'
 import { Typechecker } from '../typecheck/typechecker'
 import { createPool } from './pool'
-import type { WorkerPool } from './pool'
+import type { ProcessPool } from './pool'
 import { createBenchmarkReporters, createReporters } from './reporters/utils'
 import { StateManager } from './state'
 import { resolveConfig } from './config'
@@ -33,7 +33,7 @@ export class Vitest {
   reporters: Reporter[] = undefined!
   coverageProvider: CoverageProvider | null | undefined
   logger: Logger
-  pool: WorkerPool | undefined
+  pool: ProcessPool | undefined
   typechecker: Typechecker | undefined
 
   vitenode: ViteNodeServer = undefined!
